Refactoring and clean-up
