+++ src/extension.ts
@@ -1,24 +1,27 @@
 import * as vscode from 'vscode'
 
 import { ANG_CLI, ANG_CORE, IVersionStatus, checkForUpdate } from './file/package-manager'
+import { CheckFrequency, UpdateArgs, UpgradeChannel } from './common/enums'
 import {
-  CHECK_FREQUENCY_KEY,
   checkFrequencyExists,
   getCheckFrequency,
   getCheckFrequencyMilliseconds,
+  getCheckFrequencyPreference,
 } from './common/check-frequency.helpers'
-import { CheckFrequency, UpgradeChannel } from './common/enums'
-import { UpdateArgs, tryAngularUpdate } from './file/angular-update'
 import {
   getUpgradeChannel,
-  storeUpgradeChannel,
+  getUpgradeChannelPreference,
   upgradeChannelExists,
 } from './common/upgrade-channel.helpers'
 import {
   getVersionToSkip,
-  storeVersionToSkip,
+  getVersionToSkipPreference,
+  skipVersionCheck,
   versionToSkipExists,
 } from './common/version-to-skip.helpers'
+
+import { tryAngularUpdate } from './file/angular-update'
+import { userCancelled } from './common/common.helpers'
 
 export function activate(context: vscode.ExtensionContext) {
   console.log('Angular Evergreen is now active!')
@@ -87,10 +90,6 @@ async function runEvergreen(): Promise<void> {
   await checkForUpdates()
 }
 
-function userCancelled(userInput: string | undefined) {
-  return !userInput || userInput === '' || userInput === 'Cancel'
-}
-
 async function shouldRunEvergreen(): Promise<boolean> {
   const runEvergreenVal = await vscode.window.showInformationMessage(
     'Keep Angular evergreen?',
@@ -100,80 +99,6 @@ async function shouldRunEvergreen(): Promise<boolean> {
   )
 
   return !!runEvergreenVal && runEvergreenVal !== '' && runEvergreenVal !== 'Cancel'
-}
-
-async function getCheckFrequencyPreference(): Promise<string | undefined> {
-  const checkFrequencyVal = await vscode.window.showInformationMessage(
-    'How often would you like to check for updates (this can be changed in settings.json)?',
-    {},
-    CheckFrequency.OnLoad,
-    CheckFrequency.Hourly,
-    CheckFrequency.Daily,
-    CheckFrequency.Weekly,
-    CheckFrequency.BiWeekly
-  )
-
-  if (checkFrequencyVal && checkFrequencyVal !== '') {
-    await vscode.workspace
-      .getConfiguration()
-      .update(CHECK_FREQUENCY_KEY, checkFrequencyVal)
-  }
-
-  return checkFrequencyVal
-}
-
-async function getUpgradeChannelPreference(): Promise<string | undefined> {
-  const message = `Going forward, would you like to check for Latest releases (stable) or Next releases (risky)? This can be changed in the future in settings.json.`
-  const upgradeChannelVal = await vscode.window.showInformationMessage(
-    message,
-    {},
-    'Latest',
-    'Next',
-    'Cancel'
-  )
-
-  if (!!upgradeChannelVal && !userCancelled(upgradeChannelVal)) {
-    const shouldUpdateToNext = upgradeChannelVal.includes('Next')
-    storeUpgradeChannel(shouldUpdateToNext)
-  }
-
-  return upgradeChannelVal
-}
-
-async function getVersionToSkipPreference(): Promise<string | undefined> {
-  const cliOutdated = await checkForUpdate(ANG_CLI)
-  const coreOutdated = await checkForUpdate(ANG_CORE)
-  const shouldUpdateToNext = getUpgradeChannel() === UpgradeChannel.Next
-
-  const channelText = shouldUpdateToNext ? '"next"' : '"stable"'
-  const newCoreVersion = shouldUpdateToNext
-    ? coreOutdated.nextVersion
-    : coreOutdated.latestVersion
-  const newCliVersion = shouldUpdateToNext
-    ? cliOutdated.nextVersion
-    : cliOutdated.latestVersion
-  const versionOutdatedMsg = `New update available! One or more of your Angular packages are behind the most recent ${channelText} release. Would you like to update?
-      \r\nAngular Core: ${
-        coreOutdated.currentVersion
-      } -> ${newCoreVersion}\r\nAngular CLI: ${
-    cliOutdated.currentVersion
-  } -> ${newCliVersion}`
-
-  let versionToSkipVal = await vscode.window.showInformationMessage(
-    versionOutdatedMsg,
-    { modal: true },
-    'Update Now',
-    'Remind Me Next Release'
-  )
-
-  if (versionToSkipVal && versionToSkipVal.includes('Remind Me')) {
-    storeVersionToSkip(
-      shouldUpdateToNext ? coreOutdated.nextVersion : coreOutdated.latestVersion
-    )
-    versionToSkipVal = ''
-  }
-
-  return versionToSkipVal
 }
 
 async function checkForUpdates(): Promise<void> {
@@ -210,24 +135,5 @@ async function doAngularUpdate(
     vscode.window.showInformationMessage(`Skipping update for Angular v${versionToSkip}.`)
   } else {
     await tryAngularUpdate(shouldUpdateToNext ? [UpdateArgs.next] : [])
-  }
-}
-
-function skipVersionCheck(
-  shouldUpdateToNext: boolean,
-  versionToSkip: string | undefined,
-  coreOutdated: IVersionStatus,
-  cliOutdated: IVersionStatus
-): boolean {
-  if (shouldUpdateToNext) {
-    return versionToSkip === coreOutdated.nextVersion &&
-      versionToSkip === cliOutdated.nextVersion
-      ? true
-      : false
-  } else {
-    return versionToSkip === coreOutdated.latestVersion &&
-      versionToSkip === cliOutdated.latestVersion
-      ? true
-      : false
   }
 }
